Update README.md
Matlab scripts were used in Stowell & Wang 2025 to verify that optogenetic stimulation was effectively stimulating the mesofrontal circuit. Baseline fluorescence (F0) was defined as the average fluorescent signal (Ft) in the first 15s of the time series. Changes in the calcium signaling (ΔF/F) was calculated as (Ft-F0)/F0. We compared the average of the 10 time series to the baseline calcium activity in the ROI.